pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rh_cognitv_lite"
version = "0.0.0b0"
description = "Cognitive Skill-driven Orchestration Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "rhodie", email = "rhandref@gmail.com"},
]
dependencies = [
    "pydantic>=2.12.5",
    "jsonpatch>=1.33",
    "ulid-py>=1.1.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/rhodief/rh-cognitv_lite"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]

[tool.setuptools.package-data]
"rh_cognitv_lite" = ["py.typed"]
"rh_cognitv_lite.execution_platform" = ["py.typed"]
"rh_cognitv_lite.orchestrator" = ["py.typed"]
"rh_cognitv_lite.cognitive" = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
